+++ src/main/kotlin/io/github/ladysnake/chenille/helpers/MavenHelper.kt
@@ -54,7 +54,7 @@ internal object MavenHelper {
                             }
                         }
                     } else {
-                        println("Cannot configure Ladysnake Maven publication; please define ext.ladysnakeMavenUsername and ext.ladysnakeMavenPassword before running publish")
+                        println("Cannot configure Ladysnake Maven publication; please define ext.ladysnake_maven_username and ext.ladysnake_maven_password before running publish")
                     }
                 }
             }